+++ lib/ruta.dart
@@ -15,6 +15,7 @@ export 'package:ruta/src/core/_internal.dart'
         toShelfHandler,
         toShelfMiddleware;
 export 'package:ruta/src/core/api_exceptions.dart';
+export 'package:ruta/src/core/default_ruta_run.dart';
 export 'package:ruta/src/core/endpoint.dart';
 export 'package:ruta/src/core/handlers.dart';
 export 'package:ruta/src/core/hot_reload.dart';