+++ lib/src/core/default_ruta_run.dart
@@ -1,0 +1,11 @@
+import 'dart:io';
+
+import 'package:ruta/ruta.dart';
+
+Future<HttpServer> defaultRutaRun(
+  Handler Function() handlerCallback,
+  InternetAddress address,
+  int port,
+) {
+  return serve(handlerCallback(), address, port);
+}